[INJIMOB-1494] integrate vcverifier java library into android (24+ API)
As of June 2024, from Android API version >= 25, vc-verifier library works for vc verification. In Android API 24 version, we are facing "Trust anchor for certification path not found." issue.

Note: This commit of inji mobile uses vc-verifier library's version which is not published to maven central. To get the suitable code, please publish vc-verifier library to maven local and proceed.

Signed-off-by: KiruthikaJeyashankar <81218987+KiruthikaJeyashankar@users.noreply.github.com>